webrepl.html: Add basic file get/put functionality.
At the moment the files to get/put are hardcoded, but it works.

To test: create "test.txt" on ESP and then use "Download" button to get
it (you'll need to have JS console open to see the contents).

Press the "Upload" button to create a "test2.txt" file on the ESP with
9 "@" symbols and a newline.