webrepl.html: Deselect previously selected file before reading
Without this, if you select a file, make modifications and select the
same file again, it does not read the latest changes as the file inputs
change event does not fire. To work around, you have to select a
different file, or cancel the file dialog to clear the field value
before trying your modified file again.