webrepl: Allow the page to run from the device (over HTTP).
The device will respond to a non-WS request with a simple page that loads
websocket_content.js from a static host (http or https). However, even
if the resources are https, the page is still http and therefore allows
requesting to a WS (not WSS) websocket on the device.

Updated the warning about HTTPS and the default-ws-URL logic.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>